♻️ :: Double로 수정

+++ src/main/kotlin/com/cheajib/cheajibserver/domain/restaurant/presentation/dto/response/QueryRestaurantResponse.kt
@@ -3,7 +3,7 @@ package com.cheajib.cheajibserver.domain.restaurant.presentation.dto.response
 data class QueryRestaurantResponse(
     val restaurantName: String,
     val address: String,
-    val starPoint: Int,
+    val starPoint: Double,
     val imageUrl: String,
     val mainMenuList: String,
     val isVerify: Boolean
